- Korjaus kuljettajan tilastojen wriverWinRates alustukseen - Exception korjauksia - HipposUpdatingRegression::getWithR funktio, joka palauttaa myös R-arvon - ObserverFramework käyttämään getWithR-functiota - Value luokan parameteihin liittyviä muutoksia - ValueHorselle AccendingComparaattori valinnaiseksi

+++ src/java/hippos/math/SumReg.java
@@ -75,7 +75,7 @@ public class SumReg {
     }
 
     public int size() {
-        return xValue.getCount();
+        return xValue.getCount().intValue();
     }
 
 }
